requirements.txt
numpy
dynamixel-sdk
pyyaml